Only return methods return Optional, Comments api: create and delete

+++ src/main/java/com/example/webappspringboot/services/PostService.java
@@ -1,14 +1,11 @@
 package com.example.webappspringboot.services;
 
-import com.example.webappspringboot.Exceptions.ResourceNotFoundException;
-import com.example.webappspringboot.models.Comment;
+import com.example.webappspringboot.Exceptions.PostNotFoundException;
 import com.example.webappspringboot.models.Post;
 import com.example.webappspringboot.repositories.PostRepository;
-import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.Column;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,21 +30,40 @@ public class PostService {
         return postRepository.findById(id);
     }
 
-    public Optional<Post> updatePostById(Integer id, Post post){
+    public boolean updatePostById(Integer id, Post post){
        post.setId(id); //if I receive body without id
-       return postRepository.findById(id)
-               .map(postInDatabase -> postRepository.save(post));
-       //is it OK?
+       if(checkById(id)){
+           postRepository.save(post);
+       }
+       return true;
     }
 
-    public Optional<Post> deletePostById(Integer id) {
-        var postToDelete = postRepository.findById(id); // must to be fetch eager or Hibernate.initialize
-        postToDelete.ifPresent(post -> {
-            Hibernate.initialize(post.getComments());// PROBLEM BECAUSE IT CALLS QUERY TWO TIMES INSTEAD OF ONLY ONCE
+    public boolean deletePostById(Integer id) {
+        if(checkById(id)){ //check if it is in DB or throws an error
             postRepository.deleteById(id);
-        }); // Hibernate.initialize is used because of lazy fetch
-        return postToDelete; //MAYBE I DON'T WANT TO RETURN ANYTHING
-        //could it be written better?
+        }
+        return true;
+    }
+
+
+
+
+
+
+
+
+
+
+
+
+    public boolean checkById(Integer id){
+        var postInDatabase = postRepository.findById(id);
+        if(postInDatabase.isEmpty()){
+            throw new PostNotFoundException(id);
+        }
+        else{
+            return true;
+        }
     }
